Move domain/grid specific code to a new module.
